Fix seasonality for zero inflation curves.
The fix is checked in unit tests for an inflation index with and without
interpolation.

The fix is required so that when you take the ratio of projected
inflation fixings from a zero inflation term structure, with and without
seasonality, that you get back the input seasonalities. The unit test
was not picking this up because the tolerance of 0.01 for comparing
index fixings was too large.